Clean 'parser.c', reorganisation des src et des .h
