+++ include/utils.h
@@ -1,20 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_isspace.c                                       :+:      :+:    :+:   */
+/*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: emonacho <emonacho@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/08 21:54:36 by emonacho          #+#    #+#             */
-/*   Updated: 2025/04/18 18:05:20 by emonacho         ###   ########.fr       */
+/*   Created: 2025/04/18 17:29:45 by emonacho          #+#    #+#             */
+/*   Updated: 2025/04/18 17:57:15 by emonacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#ifndef UTILS_H
+# define UTILS_H
 
-int	ft_isspace(char c)
-{
-	if (c == 32 || (c >= 9 && c <= 13))
-		return (1);
-	return (0);
-}
+// TO REMOVE
+//////////////////////////////////
+int		debug(char *path);
+void	print_list(t_list *head);
+//////////////////////////////////
+
+// list.c
+t_list	*create_node(char *data);
+void	add_back(t_list	**head, char *data);
+void	add_front(t_list **head, char *data);
+void	free_list(t_list *head);
+
+// lexer_utils.c
+int		is_sep(char c);
+
+#endif
