+++ src/lexer/lexer_utils.c
@@ -1,23 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils.c                                            :+:      :+:    :+:   */
+/*   lexer_utils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: timmi <timmi@student.42.fr>                +#+  +:+       +#+        */
+/*   By: emonacho <emonacho@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 09:48:04 by timmi             #+#    #+#             */
-/*   Updated: 2025/04/17 11:46:45 by timmi            ###   ########.fr       */
+/*   Updated: 2025/04/18 17:48:53 by emonacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/minishell.h"
-
-int	is_space(char c)
-{
-	if (c == 32 || (c >= 9 && c <= 13))
-		return (1);
-	return (0);
-}
+#include "../../include/minishell.h"
 
 int	is_sep(char c)
 {
@@ -32,6 +25,7 @@ int	is_sep(char c)
 	}
 	return (0);
 }
+
 char	*get_quote(char *cmd, int i)
 {
 	char	quote;
@@ -75,7 +69,7 @@ char	*get_sep(char *cmd, int i)
 		if (!sep)
 			return (NULL);
 		sep[0] = cmd[i];
-		sep[1] = '\0';	
+		sep[1] = '\0';
 	}
 	return (sep);
 }
@@ -86,13 +80,13 @@ char	*get_word(char *cmd, int i)
 	char	*word;
 
 	len = 0;
-	while (cmd[i + len] && !(is_sep(cmd[i + len]) || is_space(cmd[i + len])))
+	while (cmd[i + len] && !(is_sep(cmd[i + len]) || ft_isspace(cmd[i + len])))
 		len++;
 	word = malloc((sizeof(char) * len) + 1);
 	if (!word)
 		return (NULL);
 	len = 0;
-	while (cmd[i] && !(is_sep(cmd[i]) || is_space(cmd[i])))
+	while (cmd[i] && !(is_sep(cmd[i]) || ft_isspace(cmd[i])))
 		word[len++] = cmd[i++];
 	word[len] = '\0';
 	return (word);
