+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: timmi <timmi@student.42.fr>                +#+  +:+       +#+        */
+/*   By: emonacho <emonacho@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 09:49:18 by timmi             #+#    #+#             */
-/*   Updated: 2025/04/10 16:07:36 by timmi            ###   ########.fr       */
+/*   Updated: 2025/04/18 18:04:40 by emonacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,13 +22,12 @@ void	prompt_loop(char *prompt, t_shell *s)
 	char	*line_read;
 	int		loop;
 	t_list	*head;
-	(void) s;
 
+	(void) s;
 	loop = 1;
 	while (loop)
 	{
 		line_read = readline(prompt);
-		// simple_token_interpreter(s, line_read);
 		if (line_read && *line_read)
 		// need to add a check to not print strings containing only spaces
 		{
@@ -45,6 +44,7 @@ int	main(int argc, char **argv)
 {
 	t_shell	s;
 	char	*prompt;
+
 	if (argc > 1)
 	{
 		printf("\nEntering Debug mode !\n\n");
@@ -54,6 +54,6 @@ int	main(int argc, char **argv)
 	{
 		prompt = create_prompt();
 		prompt_loop(prompt, &s);
-		free(prompt);	
+		free(prompt);
 	}
 }
